Added tree.html to help pin/share svg files in GitHub's history
Long story short, GitHub only ever serves svg files with the
image/svg+xml mimetype, which has the unfortunate side-effect of
disabling javascript/interactivity. I think this might be security
related, but not sure.

With javascript/interactivity limited to gh-pages, this makes it
difficult to pin a specific version of an svg file while keeping it user
friendly. Users can still _download_ the specific version from GitHub,
and then javascript/interactivity works, but stil...

---

To get around this, tree.html uses javascript to parse the url's "hash"
part for a commit + path, and loads it into our own html.

The modern web makes this difficult, but also surprisingly possible?
Simply referencing the target svg with <object data=(github url)> is
denied by X-Frame-Options, but using an explicit XMLHttpRequest +
encoding into a base64 data uri seems to work?

We also can't provide data uris as top-level objects, so the best we can
do with the raw link is point back to GitHub's image/svg+xml mimetype
version. At that point it's up to the user to navigate svg's weirdness.

At the very least, if everything breaks, the url still encodes the
commit hash and path. So we'll always know what the specific version of
the file was if committed/published elsewhere.