p26: Reverted ops/s instead of B/s for bench_p26_wt_many
As is, this was inconsistent between the litmus and wt benchmarks.

Adopting ops/s in litmus benchmarks is tricky, since we use n
consistently for size all the time (actually, it's probably just tedious
since we'd probably need a new result rule for n=n/CHUNK).

And ops/s vs B/s is low-priority and may make more sense as a
last-minute tweak.