Added heap/stack measurements
This uses a combination of __wrap_malloc and __builtin_frame_address to
measure stack/heap usage from inside the bench_runner itself.

Results are somewhat expected, though note we're measuring on 1. x86_64
(or whatever your host is) and not thumb (the biggest difference will be
64-bit pointers probably), and 2. asserts/logging is still enabled! But
in theory if we cross-compile, disable asserts/logs, and simulate via
QEMU the results should actually be quite accurate. This is still a TODO
(prettyasserts needs to be modified as we rely on => operator
side-effects).

We explicitly ignore heap operations outside the bench (but bench
mallocs themselves are included), and we only measure stack between
bench entry and bd ops. This may miss deep stacks that don't terminate
in a bd op, but at least for littlefs those seem rare. According to
static measurements all hot-paths terminate in lfs3_bd_read.