Fixed heap measurement underflow caused by cached spiffs buffer
We can't allow heap allocations to be created in one bench and freed in
another, it breaks our heap measurements. Hopefully just writing that
out makes sense as to why.

The escape hatch here is bench_heap_pause/resume, but this _hides_ all
heap allocations from the heap measurer, and it would be unfair to the
other filesystems to not penalize spiffs for the buffer needed to align
reads/progs.

So as a slight twist, we use bench_heap_pause/resume to hide the cached
buffer, but emulate a malloc/free call explicitly with
bench_heap_pause/resume. This should set the heap watermark
appropriately without adding the overhead of repeated malloc/free calls
to our bench runner.