Dropped block_recycles 1000 -> 100
Surprisingly, block_recycles had no measurable effect on our benchmarks,
probably due to implicit CoBW in logs.

Both 1000 and 100 are arbitrary, but 100 in theory leads to better
wear-leveling. Actually 100 is a compromise between 1000 and 10, which
may be an even better configuration. Though I can't shake the feeling
more benchmarking is needed...